improved accuracy and things, rebase this later

+++ src/libc/log1p.c
@@ -1,7 +1,12 @@
 #include <math.h>
 
+/**
+ * @remarks Minimum relative precision of:
+ * 2^-20    at +1.250000596e-01 with ideal logf
+ * 2^-19.67 at +1.250167489e-01 with current logf
+ */
 float log1pf(float x) {
-    if (fabs(x) <= 0.125) {
+    if (fabsf(x) <= 0.125f) {
         // pade(series(ln(1+x),x=0,6,polynom),x,5,3)
         // (-57*x**2-90*x)/(x**3-21*x**2-102*x-90)
         // relative error less than 1e-7
